requirements.txt
aniso8601==7.0.0
anyascii==0.1.7
asgiref==3.3.1
beautifulsoup4==4.8.2
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-cors-headers==3.5.0
django-extensions==3.1.0
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2020.6.1
openpyxl==3.0.5
Pillow==7.2.0
promise==2.3
pytz==2020.4
requests==2.25.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.26.2
wagtail==2.10.2
wagtail-grapple==0.9.2
wagtail-headless-preview==0.1.4
wagtail-nav-menus==3.3
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
